api-flask/requirements.txt
flask==0.12.2
flask_restful==0.3.6
flask_script==2.0.6
flask_migrate==2.1.1
marshmallow==2.14.0
flask_sqlalchemy==2.3.2
flask_marshmallow==0.8.0
marshmallow-sqlalchemy==0.13.2
flask-bcrypt==0.7.1
pyjwt==1.7.1
PyMySQL==0.9.3
gunicorn==20.0.4
flask-cors==3.0.8
requests==2.24.0
flask-wtf==0.14.3
Werkzeug==0.16.1
python-dotenv==0.13.0